requirements.txt
aiomysql==0.0.21
alembic==1.6.5
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
databases==0.4.3
fastapi==0.65.2
h11==0.12.0
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.1
ormar==0.10.12
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pylint==2.8.3
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-multipart==0.0.5
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.3.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.5
uvicorn==0.14.0
wrapt==1.12.1
